pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coffeecode-uploader"
version = "2.0.0"
description = "Easy Python upload manager for images, files and media. Python fork of robsonvleite/uploader (PHP)."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "Kaue Leal", email = "kaue.lima@totvs.com.br" },
]
maintainers = [
    { name = "Kaue Leal", email = "kaue.lima@totvs.com.br" },
]
keywords = ["upload", "image upload", "file upload", "media upload", "coffeecode"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://github.com/kaueleal/uploader"
"Original PHP" = "https://github.com/robsonvleite/uploader"
Issues = "https://github.com/kaueleal/uploader/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/coffeecode_uploader"]

[tool.pytest.ini_options]
testpaths = ["tests"]
